requirements.txt
# Flask and Extensions
Flask==3.0.2
Flask-Caching==2.1.0
python-dotenv==1.0.1

# Google API
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0

# OpenAI
openai==1.12.0

# Utilities
requests==2.31.0
Werkzeug==3.0.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5 
gunicorn